services/ml-tabular-worker/pyproject.toml
[project]
name = "ml-tabular-worker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["fastapi>=0.115.0", "uvicorn>=0.32.0", "pyyaml>=6.0.2"]
